RavenDB-18216 implemented /databases/*/studio/index-type
